some bug-fixes
- changed theme_id to Integer according to the database model
- some fixes in tests

+++ src/main/java/com/nasnav/dto/OrganizationDTO.java
@@ -33,7 +33,7 @@ public class OrganizationDTO {
 
         @ApiModelProperty(value = "theme id used in the organization", example = "123")
         @JsonProperty("theme_id")
-        public Long themeId;
+        public Integer themeId;
 
         @ApiModelProperty(value = "URL to twitter account", example = "https://www.twitter.com/fortunestores")
         @JsonProperty("social_twitter")
